+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - **feature:** optimize the default config schema.
 - **fix:** make it working in MC version [1.20, 1.20.2].
 - **feature:** Pre-compile the regex pattern to improve performance.
+- **feature:** New `CANCEL_WITH_FAILURE` command advice type.